Re-Added Support for CoFH Mods (WIP)
The previous textures available for 1.7.10 of CoFH Mods has been updated
to work in the 1.12 mc version of the mods. There's still a large amount
of the assets missing so don't expect everything to be there. Not to
mention some of the textures are ever so slightly incorrect or wrong.
The textures will be fixed and more of them will be added in due time.

- ADDED: thermalfoundation textures
- ADDED: thermalexpansion textures