Renamed stone variants
In 1.13, Mojang have renamed the stone variants to be simply the variant
itself (by removing the stone_ prefix). Also, polished (prev. known as
smooth) variants have been prefixed with polished.

There is however a caveat. In the most up-to-date snapshot (18w07c)
granite has yet to be renamed (the polished one was but not the regular
granite). This should be reported to Mojang if its not already fixed by
release otherwise it will be sure to drive everyone up the wall, me
included.

(I am unable to report it myself because I cannot gain access to my
account.)

Signed-off-by: SnowShock35 <xsnowshockx35@gmail.com>