Updated Spashes.txt With Love
I added a single line to the 'splashes.txt' file in the Minecraft assets folder. It was just a single line... but a significant line. I felt that the document was lacking a beautiful §4 to line 587. Line 587... not 586... not 588... but 587. It is deserving of the same §4 characters as all of the others (except 588 which is empty). I made this change, not out of the urge to be extremely sarcastic but to teach myself how each commit should be done, properly and with purpose. You might think this is overboard for such a simple commit but I felt very strongly that this needed to be done. Enjoy your new §4 and don't take it for granted. Yours truly, CyanideX.